Caching Test 6.3
